Enhance WgetTest with additional test cases for error handling and response limits
