+++ src/test/java/net/openhft/chronicle/core/io/WgetTest.java
@@ -4,120 +4,245 @@ import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.nio.charset.StandardCharsets;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import org.junit.jupiter.api.Test;
+
+import java.io.*;
+import java.net.URL;
+import java.net.URLConnection;
+import java.nio.charset.StandardCharsets;
+import java.util.concurrent.*;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * Tests for the Wget class, which provides a simple way to fetch content from URLs.
+ */
 class WgetTest {
-
-    /* -------------------------------------------------
-     * Helpers
-     * ------------------------------------------------- */
-
-    /** Returns a Wget instance whose “HTTP response” is a fixed byte array. */
-    private static Wget stubbed(String body) {
-        return new Wget.Builder()
-                .connectionProvider(u ->
-                        new ByteArrayInputStream(body.getBytes(StandardCharsets.UTF_8)))
-                .build();
-    }
-
-    /* -------------------------------------------------
-     * Happy-path behaviour
-     * ------------------------------------------------- */
 
     @Test
     void fetch_appends_response_body() throws IOException {
-        StringBuilder sb = new StringBuilder();
-
-        stubbed("hello world").fetch("http://does.not.matter", sb);
-
-        assertEquals("hello world", sb.toString());
-    }
-
-    /* -------------------------------------------------
-     * CharsetDetector is consulted
-     * ------------------------------------------------- */
-
-    @Test
-    void custom_charset_detector_is_used() throws IOException {
-        // Body encoded in ISO-8859-1 so the final byte sequence is different to UTF-8.
-        String original = "Café";
-        byte[] isoBytes = original.getBytes("ISO-8859-1");
-
+        String expected = "hello world";
         Wget wget = new Wget.Builder()
-                .connectionProvider(u -> new ByteArrayInputStream(isoBytes))
-                .charsetDetector((in, ct) -> java.nio.charset.StandardCharsets.ISO_8859_1)
+                .connectionProvider(u -> new ByteArrayInputStream(expected.getBytes(StandardCharsets.UTF_8)))
                 .build();
-
         StringBuilder sb = new StringBuilder();
-        wget.fetch("http://anything", sb);
-
-        assertEquals(original, sb.toString());
+        wget.fetch("http://does.not.matter", sb);
+        assertEquals(expected, sb.toString());
     }
-
-    /* -------------------------------------------------
-     * Size limits
-     * ------------------------------------------------- */
-
-    @Test
-    void oversized_body_triggers_IOException() {
-        byte[] tenBytes = new byte[10];
-
-        Wget wget = new Wget.Builder()
-                .connectionProvider(u -> new ByteArrayInputStream(tenBytes))
-                .maxResponseBytes(5)          // ridiculously small
-                .build();
-
-        assertThrows(IOException.class,
-                () -> wget.fetch("http://x", new StringBuilder()),
-                "Size limit exceeded");
-    }
-
-    /* -------------------------------------------------
-     * URL validation
-     * ------------------------------------------------- */
 
     @Test
     void invalid_scheme_throws_MalformedURLException() {
-        Wget wget = stubbed("");     // CP will never be called
-
-        assertThrows(MalformedURLException.class,
-                () -> wget.fetch("ftp://example.com", new StringBuilder()));
+        Wget wget = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream(new byte[0]))
+                .build();
+        assertThrows(IOException.class, () -> wget.fetch("ftp://example.com", new StringBuilder()));
     }
 
-    /* -------------------------------------------------
-     * Null-safety
-     * ------------------------------------------------- */
-
     @Test
-    void null_appendable_throws_NullPointerException() {
-        Wget wget = stubbed("");
-
-        assertThrows(NullPointerException.class,
-                () -> wget.fetch("http://x", null));
+    void null_appendable_throws() {
+        Wget wget = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream(new byte[0]))
+                .build();
+        assertThrows(NullPointerException.class, () -> wget.fetch("http://x", null));
     }
 
-    /* -------------------------------------------------
-     * Collaborator interaction
-     * ------------------------------------------------- */
 
     @Test
-    void connectionProvider_receives_the_same_URL() throws IOException {
-        AtomicReference<URL> seen = new AtomicReference<>();
+    void body_equal_to_limit_is_allowed() throws IOException {
+        byte[] five = "12345".getBytes(StandardCharsets.UTF_8);
+        Wget wget = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream(five))
+                .maxResponseBytes(5)
+                .build();
+        StringBuilder sb = new StringBuilder();
+        wget.fetch("http://x", sb);
+        assertEquals("12345", sb.toString());
+    }
 
+    @Test
+    void unknown_content_length_still_enforced() {
+        InputStream neverEnding = new InputStream() {
+            @Override
+            public int read() {
+                return 'A';
+            }
+        };
+        Wget wget = new Wget.Builder()
+                .connectionProvider(u -> neverEnding)
+                .maxResponseBytes(128)
+                .build();
+        assertThrows(IOException.class, () -> wget.fetch("http://x", new StringBuilder()));
+    }
+
+    @Test
+    void IOException_from_connection_provider_bubbles_up() {
         Wget wget = new Wget.Builder()
                 .connectionProvider(u -> {
-                    seen.set(u);
-                    return new ByteArrayInputStream("ok".getBytes(StandardCharsets.UTF_8));
+                    throw new IOException("boom");
                 })
                 .build();
+        assertThrows(IOException.class, () -> wget.fetch("http://x", new StringBuilder()));
+    }
 
-        wget.fetch("http://example.com/resource", new StringBuilder());
+    @Test
+    void null_charset_detector_result_falls_back_to_utf8() throws IOException {
+        byte[] café = "Café".getBytes(StandardCharsets.UTF_8);
+        Wget wget = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream(café))
+                .charsetDetector((in, ct) -> null)
+                .build();
+        StringBuilder sb = new StringBuilder();
+        wget.fetch("http://x", sb);
+        assertEquals("Café", sb.toString());
+    }
 
-        assertEquals("http://example.com/resource", seen.get().toString());
+    @Test
+    void appendable_exception_is_propagated() {
+        Appendable broken = new Appendable() {
+            @Override
+            public Appendable append(char c) throws IOException {
+                throw new IOException("disk full");
+            }
+
+            @Override
+            public Appendable append(CharSequence csq) {
+                return this;
+            }
+
+            @Override
+            public Appendable append(CharSequence csq, int s, int e) {
+                return this;
+            }
+        };
+        Wget wget = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream("x".getBytes()))
+                .build();
+        assertThrows(IOException.class, () -> wget.fetch("http://x", broken));
+    }
+
+    @Test
+    void static_url_method_rejects_oversized_url() {
+        StringBuilder sbUrl = new StringBuilder("http://a");
+        while (sbUrl.length() <= 2100) sbUrl.append('b');
+        assertThrows(IllegalArgumentException.class, () -> Wget.url(sbUrl.toString(), new StringBuilder()));
+    }
+
+    @Test
+    void fetch_is_thread_safe_when_instance_is_shared() throws Exception {
+        Wget wget = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream("ok".getBytes()))
+                .build();
+        ExecutorService pool = Executors.newFixedThreadPool(4);
+        AtomicInteger successes = new AtomicInteger();
+        Callable<Void> task = () -> {
+            StringBuilder sb = new StringBuilder();
+            wget.fetch("http://x", sb);
+            if ("ok".contentEquals(sb)) successes.incrementAndGet();
+            return null;
+        };
+        for (int i = 0; i < 20; i++) pool.submit(task);
+        pool.shutdown();
+        assertTrue(pool.awaitTermination(2, TimeUnit.SECONDS));
+        assertEquals(20, successes.get());
+    }
+
+    @Test
+    void limited_stream_behaves_like_eof_after_budget() throws IOException {
+        byte[] data = "abc".getBytes();
+        LimitedInputStream lim = new LimitedInputStream(new ByteArrayInputStream(data), 3);
+        ByteArrayOutputStream copy = new ByteArrayOutputStream();
+        for (int b; (b = lim.read()) != -1; ) copy.write(b);
+        assertArrayEquals(data, copy.toByteArray());
+        assertEquals(-1, lim.read());
+    }
+
+    @Test
+    void zero_budget_allows_empty_body_but_blocks_data() throws IOException {
+        Wget empty = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream(new byte[0]))
+                .maxResponseBytes(0)
+                .build();
+        StringBuilder sb = new StringBuilder();
+        empty.fetch("http://x", sb);
+        assertEquals("", sb.toString());
+
+        Wget tooMuch = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream("x".getBytes()))
+                .maxResponseBytes(0)
+                .build();
+        assertThrows(IOException.class, () -> tooMuch.fetch("http://x", new StringBuilder()));
+    }
+
+    @Test
+    void negative_budget_rejected() {
+        assertThrows(IllegalArgumentException.class, () -> new Wget.Builder().maxResponseBytes(-1).build());
+    }
+
+    @Test
+    void limited_stream_byte_array_path_respects_limit() throws IOException {
+        byte[] data = "abcdef".getBytes(StandardCharsets.UTF_8);
+        LimitedInputStream lim = new LimitedInputStream(new ByteArrayInputStream(data), 4);
+        byte[] buf = new byte[10];
+        int n = lim.read(buf);
+        assertEquals(4, n);
+        assertEquals("abcd", new String(buf, 0, n, StandardCharsets.UTF_8));
+        assertThrows(IOException.class, () -> lim.read(buf));
+    }
+
+    @Test
+    void charset_detector_exception_bubbles_up() {
+        Wget wget = new Wget.Builder()
+                .connectionProvider(u -> new ByteArrayInputStream("x".getBytes()))
+                .charsetDetector((in, ct) -> {
+                    throw new RuntimeException("boom");
+                })
+                .build();
+        assertThrows(RuntimeException.class, () -> wget.fetch("http://x", new StringBuilder()));
+    }
+
+    @Test
+    void default_provider_sets_timeouts() throws Exception {
+        AtomicInteger seenConnect = new AtomicInteger(-1);
+        AtomicInteger seenRead = new AtomicInteger(-1);
+
+        URL synthetic = new URL(null, "http://timeout.test", new java.net.URLStreamHandler() {
+            @Override
+            protected URLConnection openConnection(URL u) {
+                return new URLConnection(u) {
+                    @Override
+                    public void connect() {
+                    }
+
+                    @Override
+                    public void setConnectTimeout(int v) {
+                        seenConnect.set(v);
+                    }
+
+                    @Override
+                    public void setReadTimeout(int v) {
+                        seenRead.set(v);
+                    }
+
+                    @Override
+                    public InputStream getInputStream() {
+                        return new ByteArrayInputStream(new byte[0]);
+                    }
+                };
+            }
+        });
+
+        int ct = 1_234, rt = 5_678;
+        Wget wget = new Wget.Builder().connectTimeoutMs(ct).readTimeoutMs(rt).build();
+        java.lang.reflect.Field f = Wget.class.getDeclaredField("connectionProvider");
+        f.setAccessible(true);
+        Wget.ConnectionProvider provider = (Wget.ConnectionProvider) f.get(wget);
+        provider.open(synthetic).close();
+        assertEquals(ct, seenConnect.get());
+        assertEquals(rt, seenRead.get());
     }
 }
